# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * obtv-ai media intelligence platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MediaAsset {
  id: string;
  filename: string;
  /** @nullable */
  original_path?: string | null;
  /** @nullable */
  proxy_path?: string | null;
  /** @nullable */
  thumbnail_url?: string | null;
  /** @nullable */
  duration_seconds?: number | null;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  fps?: number | null;
  /** @nullable */
  codec?: string | null;
  /** @nullable */
  file_size_bytes?: number | null;
  /** pending | processing | ready | error */
  status: string;
  /**
     * Current processing stage name
     * @nullable
     */
  processing_stage?: string | null;
  /**
     * 0-100 percent
     * @nullable
     */
  processing_progress?: number | null;
  /** @nullable */
  scene_count?: number | null;
  /** @nullable */
  speaker_count?: number | null;
  created_at: string;
  /** @nullable */
  updated_at?: string | null;
}

export interface MediaListResponse {
  items: MediaAsset[];
  total: number;
}

export interface MediaIngestInput {
  /** Absolute path to the source media file on the server */
  file_path: string;
  title?: string;
}

export type LibraryStatsStatusCounts = {[key: string]: number};

export interface LibraryStats {
  total_assets: number;
  total_duration_seconds: number;
  status_counts: LibraryStatsStatusCounts;
  storage_bytes: number;
  recent_activity?: MediaAsset[];
}

export interface Scene {
  id: string;
  media_id: string;
  /** Seconds */
  start_time: number;
  /** Seconds */
  end_time: number;
  /** @nullable */
  thumbnail_url?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  embedding_id?: string | null;
}

export interface TranscriptSegment {
  id: string;
  media_id: string;
  start_time: number;
  end_time: number;
  text: string;
  /** @nullable */
  speaker?: string | null;
  /** @nullable */
  confidence?: number | null;
}

export type FaceClusterAppearancesItem = {
  start_time: number;
  end_time: number;
};

export interface FaceCluster {
  cluster_id: string;
  media_id: string;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  thumbnail_url?: string | null;
  appearances: FaceClusterAppearancesItem[];
}

export interface SearchQuery {
  query: string;
  /**
     * Restrict search to a specific asset
     * @nullable
     */
  media_id?: string | null;
  /** transcript | visual | combined */
  search_type?: string;
  limit?: number;
}

export interface SearchResult {
  media_id: string;
  filename: string;
  /** @nullable */
  thumbnail_url?: string | null;
  start_time: number;
  end_time: number;
  score: number;
  /** transcript | visual */
  match_type: string;
  /**
     * Matching transcript text
     * @nullable
     */
  snippet?: string | null;
}

export interface SearchResponse {
  results: SearchResult[];
  query: string;
  took_ms: number;
}

export interface SearchHistoryItem {
  id: string;
  query: string;
  result_count: number;
  searched_at: string;
}

export interface ProcessingJob {
  id: string;
  media_id: string;
  /** @nullable */
  filename?: string | null;
  /** ingest | proxy | audio_extract | transcribe | diarize | scene_detect | visual_embed | face_detect | index */
  job_type: string;
  /** pending | running | success | error | cancelled */
  status: string;
  /**
     * 0-100
     * @nullable
     */
  progress?: number | null;
  /** @nullable */
  error_message?: string | null;
  logs?: string[];
  retry_count?: number;
  created_at: string;
  /** @nullable */
  started_at?: string | null;
  /** @nullable */
  finished_at?: string | null;
}

export interface AIQuestion {
  question: string;
  /** @nullable */
  conversation_id?: string | null;
  /** @nullable */
  media_id?: string | null;
}

export interface AICitation {
  media_id: string;
  filename: string;
  start_time: number;
  end_time: number;
  /** @nullable */
  snippet?: string | null;
}

export interface AIAnswer {
  answer: string;
  conversation_id: string;
  citations: AICitation[];
}

export interface Conversation {
  id: string;
  /** @nullable */
  title?: string | null;
  created_at: string;
  message_count: number;
}

export interface Clip {
  id: string;
  media_id: string;
  /** @nullable */
  filename?: string | null;
  start_time: number;
  end_time: number;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface ClipList {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  created_at: string;
  clips: Clip[];
}

export type ClipListInputClipsItem = {
  media_id: string;
  start_time: number;
  end_time: number;
  label?: string;
};

export interface ClipListInput {
  name: string;
  description?: string;
  clips?: ClipListInputClipsItem[];
}

export type ClipListUpdateClipsItem = {
  media_id: string;
  start_time: number;
  end_time: number;
  label?: string;
};

export interface ClipListUpdate {
  name?: string;
  description?: string;
  clips?: ClipListUpdateClipsItem[];
}

export interface ClipExportInput {
  /** edl | csv | json */
  format: string;
}

export interface ClipExportResult {
  format: string;
  content: string;
  filename?: string;
}

export type ListMediaParams = {
status?: string;
limit?: number;
offset?: number;
};

export type ListJobsParams = {
media_id?: string;
status?: string;
limit?: number;
};