# services/api/app/routers/jobs.py
from typing import Optional
from fastapi import APIRouter, Depends, HTTPException, Query
from sqlalchemy.ext.asyncio import AsyncSession
from sqlalchemy import select, desc
from ..database import get_db
from ..models import ProcessingJob, MediaAsset
from ..schemas import ProcessingJobOut

router = APIRouter(prefix="/jobs", tags=["jobs"])


def _enrich(job: ProcessingJob, asset: MediaAsset | None) -> ProcessingJobOut:
    out = ProcessingJobOut.model_validate(job)
    if asset:
        out.filename = asset.filename
    return out


@router.get("", response_model=list[ProcessingJobOut])
async def list_jobs(
    media_id: Optional[str] = Query(None),
    status: Optional[str] = Query(None),
    limit: int = Query(100, ge=1, le=500),
    db: AsyncSession = Depends(get_db),
):
    q = (
        select(ProcessingJob, MediaAsset)
        .outerjoin(MediaAsset, ProcessingJob.media_id == MediaAsset.id)
        .order_by(desc(ProcessingJob.created_at))
    )
    if media_id:
        q = q.where(ProcessingJob.media_id == media_id)
    if status:
        q = q.where(ProcessingJob.status == status)
    q = q.limit(limit)
    rows = (await db.execute(q)).all()
    return [_enrich(job, asset) for job, asset in rows]


@router.get("/{id}", response_model=ProcessingJobOut)
async def get_job(id: str, db: AsyncSession = Depends(get_db)):
    row = (
        await db.execute(
            select(ProcessingJob, MediaAsset)
            .outerjoin(MediaAsset, ProcessingJob.media_id == MediaAsset.id)
            .where(ProcessingJob.id == id)
        )
    ).first()
    if not row:
        raise HTTPException(status_code=404, detail="Job not found")
    job, asset = row
    return _enrich(job, asset)


@router.post("/{id}/retry", response_model=ProcessingJobOut, status_code=202)
async def retry_job(id: str, db: AsyncSession = Depends(get_db)):
    row = (
        await db.execute(
            select(ProcessingJob, MediaAsset)
            .outerjoin(MediaAsset, ProcessingJob.media_id == MediaAsset.id)
            .where(ProcessingJob.id == id)
        )
    ).first()
    if not row:
        raise HTTPException(status_code=404, detail="Job not found")
    job, asset = row
    if job.status not in ("error", "cancelled"):
        raise HTTPException(status_code=400, detail="Only failed or cancelled jobs can be retried")

    job.status = "pending"
    job.error_message = None
    job.retry_count += 1
    job.started_at = None
    job.finished_at = None
    await db.commit()
    await db.refresh(job)

    from ..worker_client import enqueue_job
    await enqueue_job(job.job_type, job.media_id, job.id)

    return _enrich(job, asset)


@router.post("/{id}/cancel", response_model=ProcessingJobOut)
async def cancel_job(id: str, db: AsyncSession = Depends(get_db)):
    row = (
        await db.execute(
            select(ProcessingJob, MediaAsset)
            .outerjoin(MediaAsset, ProcessingJob.media_id == MediaAsset.id)
            .where(ProcessingJob.id == id)
        )
    ).first()
    if not row:
        raise HTTPException(status_code=404, detail="Job not found")
    job, asset = row
    if job.status in ("success", "error"):
        raise HTTPException(status_code=400, detail="Completed jobs cannot be cancelled")

    if job.celery_task_id:
        try:
            from ..services.celery_app import celery_app
            celery_app.control.revoke(job.celery_task_id, terminate=True)
        except Exception:
            pass

    job.status = "cancelled"
    await db.commit()
    await db.refresh(job)
    return _enrich(job, asset)